requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.24.0
qdrant-client>=1.7.0
python-dotenv==1.2.1